requirements.txt
django==2.1
unicode-slugify==0.1.3
